chore: Moved dataclasses to their own files
