+++ koswat/cost_report/summary/koswat_summary_location_matrix_builder.py
@@ -9,7 +9,7 @@ from koswat.cost_report.multi_location_profile.multi_location_profile_cost_repor
 from koswat.dike.surroundings.point.point_surroundings import PointSurroundings
 from koswat.strategies.strategy_input import (
     StrategyLocationInput,
-    StrategyLocationReinforcementCosts,
+    StrategyReinforcementTypeCosts,
 )
 
 
@@ -30,11 +30,11 @@ class KoswatSummaryLocationMatrixBuilder(BuilderProtocol):
 
     def _get_multi_location_profile_to_dict_matrix(
         self, locations_profile: MultiLocationProfileCostReport
-    ) -> dict[PointSurroundings, StrategyLocationReinforcementCosts]:
+    ) -> dict[PointSurroundings, StrategyReinforcementTypeCosts]:
         _infra_costs = locations_profile.get_infra_costs_per_location()
 
         def create_strategy_location_reinforcement_costs():
-            return StrategyLocationReinforcementCosts(
+            return StrategyReinforcementTypeCosts(
                 reinforcement_type=type(
                     locations_profile.profile_cost_report.reinforced_profile
                 ),
@@ -50,7 +50,7 @@ class KoswatSummaryLocationMatrixBuilder(BuilderProtocol):
 
     def _get_list_summary_matrix_for_locations_with_reinforcements(
         self,
-    ) -> list[dict[PointSurroundings, StrategyLocationReinforcementCosts]]:
+    ) -> list[dict[PointSurroundings, StrategyReinforcementTypeCosts]]:
         return list(
             map(
                 self._get_multi_location_profile_to_dict_matrix,
@@ -84,9 +84,7 @@ class KoswatSummaryLocationMatrixBuilder(BuilderProtocol):
 
         # 4. Sort matrix by traject order for normalized usage in Koswat.
         def to_strategy_location(
-            matrix_tuple: tuple[
-                PointSurroundings, list[StrategyLocationReinforcementCosts]
-            ]
+            matrix_tuple: tuple[PointSurroundings, list[StrategyReinforcementTypeCosts]]
         ) -> StrategyLocationInput:
             return StrategyLocationInput(
                 point_surrounding=matrix_tuple[0],
